properly handling squirrel event for windows
